% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_tests.R
\name{aggregate_tests}
\alias{aggregate_tests}
\title{Aggregate tests by a grouping variable}
\usage{
aggregate_tests(data, by = "Created Date")
}
\arguments{
\item{data}{A [data.frame()] or [tibble()].}

\item{by}{A string matching one of the column names of `data`; the default is  `"Created Date"`.}
}
\value{
an object of the same type as `data`, with the following columns added:

* `n tests`
* `n with results`
* `n positive`
* `% positive`
* `period`

The output will be arranged in ascending order of the variable identified by the `by` argument; that is, it will be arranged in chronological order of `Created Date`, by default.
}
\description{
Counts test results by a grouping variable (`Created Date`, by default) and
adds some useful columns
}
