% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal sounds sentences}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string containing the name of the animal}

\item{sound}{A character string containing the sound the animal makes}
}
\value{
A sentences matching animals with their sounds
}
\description{
Animal sounds sentences
}
\examples{
sound <- animal_sounds("lion","roar")
}
