\name{api}
\alias{api}
\title{API function}
\usage{
api(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=middle]{middle}}
  \describe{
    \item{\code{arg1}}{a parameter argument}
    \item{\code{...}}{Arguments passed on from \code{\link[=middle]{middle}} to \code{\link[=core]{core}}}
  }}
}
\description{
API function
}
