% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_lbls.R
\name{box_lbls}
\alias{box_lbls}
\title{Summary Labels for Boxplots}
\usage{
box_lbls(x, color = NULL)
}
\arguments{
\item{x}{A numeric vector for specifying data from which the boxplots are to be produced}

\item{color}{The color of the body of the boxplot. By default, the boxplot color is the same as the background color.}
}
\description{
This function labels summary statistics on an R boxplot. The summary statistics reported
are the minimum, maximum, Q1, Q3 and median. To use the function, you must specify a vector \code{x}.
}
\examples{
box_lbls(rnorm(1000), color = "pink")

x <- c(1,6,3,9,7,5,9,16,4)
box_lbls(x, color = "lightblue")

}
\author{
Samantha Bothwell \email{samantha.bothwell@cuanschutz.edu}
}
