% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_with_backticked_arg_names.R
\name{function_with_backticked_arg_names}
\alias{function_with_backticked_arg_names}
\title{A function with argument names defined using backticks}
\usage{
function_with_backticked_arg_names(`arg 1` = 1, `arg 2` = "apple")
}
\arguments{
\item{`arg}{2` a string}
}
\value{
the string \code{"goodbye"}
}
\description{
More information about this function.
}
\examples{
function_with_backticked_arg_names()
}
\keyword{internal}
