% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbox_lbls.numeric.R
\name{ggbox_lbls.numeric}
\alias{ggbox_lbls.numeric}
\title{Summary Labels for Boxplots in GGplot for class numeric}
\usage{
\method{ggbox_lbls}{numeric}(obj, color = NULL, ...)
}
\arguments{
\item{obj}{numeric object to be plotted}

\item{color}{The color of the body of the boxplot. By default, the boxplot color is the same as the background color.}

\item{...}{additional arguments passed to plotting functions}
}
\description{
This function labels summary statistics on a ggplot boxplot. The summary statistics reported
are the minimum, maximum, Q1, Q3 and median. To use the function, you must specify a vector \code{x}.
}
\examples{
ggbox_lbls(rnorm(1000), color = "pink")

x <- c(1,6,3,9,7,5,9,16,4)
ggbox_lbls(x, color = "steelblue1")

}
\author{
Samantha Bothwell \email{samantha.bothwell@cuanschutz.edu}
}
