% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbox_lbls.R
\name{ggbox_lbls}
\alias{ggbox_lbls}
\title{Summary Labels for Boxplots in GGplot}
\usage{
ggbox_lbls(obj, ...)
}
\arguments{
\item{obj}{object to be plotted. Currently, allowed data types are \code{samantha} and \code{numeric}.}

\item{...}{additional arguments passed to plotting functions}
}
\description{
This function labels summary statistics on a ggplot boxplot. The summary statistics reported
are the minimum, maximum, Q1, Q3 and median. To use the function, you must specify a vector \code{x}.
}
\examples{

\dontrun{

ggbox_lbls(rnorm(1000), color = "pink")

x <- c(1,6,3,9,7,5,9,16,4)
ggbox_lbls(x, color = "steelblue1")

}

}
\author{
Samantha Bothwell \email{samantha.bothwell@cuanschutz.edu}
}
