% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit.dot.params.R
\name{inherit.dot.params}
\alias{inherit.dot.params}
\title{Title}
\usage{
inherit.dot.params(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=aggregate_tests]{aggregate_tests}}
  \describe{
    \item{\code{data}}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[=tibble]{tibble()}}.}
    \item{\code{by}}{A string matching one of the column names of \code{data}; the default is  \code{"Created Date"}.}
  }}
}
\description{
Title
}
