% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit_params.R
\name{inherit_params}
\alias{inherit_params}
\title{Title}
\usage{
inherit_params(data, by = "Created Date")
}
\arguments{
\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[=tibble]{tibble()}}.}

\item{by}{A string matching one of the column names of \code{data};
the default is  \code{"Created Date"}.}
}
\description{
Title
}
