% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobar.R
\name{middle}
\alias{middle}
\title{intermediate function}
\usage{
middle(arg1, ...)
}
\arguments{
\item{arg1}{a parameter argument}

\item{...}{
  Arguments passed on to \code{\link[=core]{core}}
  \describe{
    \item{\code{arg2}}{another parameter argument}
  }}
}
\description{
intermediate function
}
