% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donut_univariadas.R
\name{plot_donut_univariadas_test2}
\alias{plot_donut_univariadas_test2}
\title{Generate Univariate Donut Plot}
\usage{
plot_donut_univariadas_test2(
  .base,
  .var = NULL,
  .palette = NULL,
  .hsize = 1.3,
  .plot_text_size = 6,
  .plot_text_color = "#252525",
  .legend_text_size = 16,
  .gray_category = NULL,
  .wrap_length = NULL,
  .text_threshold = 0.05,
  .omit_label = NULL
)
}
\arguments{
\item{.base}{A raw dataframe or processed one by tablas_univariadas.}

\item{.var}{Optional var to plot.}

\item{.palette}{Optional color palette.}

\item{.hsize}{Parameter that allows to modify the size of the donut.}

\item{.plot_text_size}{Parameter that allows to modify the size of the text inside the plot.}

\item{.plot_text_color}{Parameter that allows to modify the color of the text inside the plot.}

\item{.legend_text_size}{Parameter that allows to modify the size of the text inside the legend.}

\item{.gray_category}{Optional categories to be filled with gray color.}

\item{.wrap_length}{Optional parameter to define the maximum length of category labels before wrapping.}

\item{.text_threshold}{Optional parameter to define the threshold of the minimum percent to display.}

\item{.omit_label}{Optional parameter to define custom categories to omit percent to display.}
}
\value{
A plot with counts and percentages for each category of the selected variable.
}
\description{
This function generates a univariate donut plot, displaying the count
and percentage of each category of a variable within a dataframe.
It supports optional color for plot and size for plot and legend text.
}
