% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulados_banner.R
\name{tabulados_banner}
\alias{tabulados_banner}
\title{Generate Crosstabulations with Banner using expss}
\usage{
tabulados_banner(.base, .var, .banner = NULL, .wt = NULL, .cases = FALSE)
}
\arguments{
\item{.base}{A dataframe to be used for the tabulations.}

\item{.var}{A selection of variables to tabulate (can be one or more).}

\item{.banner}{Optional variable(s) to use as column banner in the tabulation.}

\item{.wt}{Optional variable to be used for weights.}

\item{.cases}{Logical. If \code{TRUE}, includes case counts in the tabulation.}
}
\value{
A list of tabulation tables, one per variable selected.
}
\description{
This function generates a list of crosstabulations using the expss package. It supports optional weighting,
optional case count inclusion, and flexible bannering.
}
