% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.package-package.R
\docType{package}
\name{test.package-package}
\alias{test.package}
\alias{test.package-package}
\title{test.package: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/test.package}
  \item \url{https://d-morrison.github.io/test.package/}
  \item Report bugs at \url{https://github.com/d-morrison/test.package/issues}
}

}
\keyword{internal}
