% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{test_function}
\alias{test_function}
\title{Add two numbers}
\usage{
test_function(a, b)
}
\arguments{
\item{a}{The first number}

\item{b}{The second number}
}
\value{
Sum of the two numbers
}
\description{
Add two numbers
}
\examples{
a = 1; b = 2; test_function(1,2)
}
