% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_tests.R
\name{aggregate_tests}
\alias{aggregate_tests}
\title{Aggregate tests by a grouping variable}
\usage{
aggregate_tests(data, by = "Created Date")
}
\arguments{
\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[=tibble]{tibble()}}.}

\item{by}{A string matching one of the column names of \code{data};
the default is  \code{"Created Date"}.}
}
\value{
an object of the same type as \code{data},
with the following columns added:
\itemize{
\item \verb{n tests}
\item \verb{n with results}
\item \verb{n positive}
\item \verb{\% positive}
\item \code{period}
}

The output will be arranged in ascending order of the variable
identified by the \code{by} argument;
that is, it will be arranged in chronological order of \verb{Created Date},
by default.
}
\description{
Counts test results by a grouping variable (\verb{Created Date}, by default) and
adds some useful columns
}
