% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_resp_multiple.R
\name{tablas_resp_multiple}
\alias{tablas_resp_multiple}
\title{Generate Frequency Tables from Multiple Response variables}
\usage{
tablas_resp_multiple(
  .base,
  .vars,
  .wt = NULL,
  na.rm = TRUE,
  .pct_formula = c("responses", "cases")
)
}
\arguments{
\item{.base}{A dataframe.}

\item{.vars}{A set of columns.}

\item{.wt}{Optional weights.}

\item{na.rm}{Logical, whether to remove NAs.}

\item{.pct_formula}{Charater, to define the type of denominator for the percentage.}
}
\value{
A tibble with counts and percentages.
}
\description{
This function generates a multiple response frequency table for a set of variables,
displaying the count and percentage of each category within the dataframe.
It supports optional weighting and can remove missing values if specified.
}
