% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_resp_multiple.R
\name{tablas_resp_multiple}
\alias{tablas_resp_multiple}
\title{Generación de frecuencias univariadas}
\usage{
tablas_resp_multiple(.base, .vars, .wt = NULL, na.rm = TRUE)
}
\arguments{
\item{.base}{A dataframe.}

\item{.vars}{Set of columns.}

\item{.wt}{Optional weights.}

\item{na.rm}{Logical, whether to remove NAs.}
}
\value{
A tibble with counts and percentages.
}
\description{
Description of what it does
}
